проверяем стили главного меню
