+++ package.json
@@ -28,6 +28,7 @@
     "stylelint": "stylelint \"source/less/**/*.less\" --syntax less",
     "test": "npm run editorconfig && npm run stylelint",
     "build": "gulp build",
+    "build-source": "gulp stylesSource",
     "start": "gulp"
   },
   "browserslist": [
